requirements.txt
Django==2.0.2
Pillow==5.0.0
django-debug-toolbar==1.9.1
django-geoposition==0.3.0
django-imagekit==4.0.2
django-simple-captcha==0.5.6
django-tinymce==2.7.0
